Reenable spell checker (#150)
Enables the spell checker CI action again with a revised workflow. This
now uses `pull_request_target` for the action context, so that PRs from
forks can have comments posted on the PR.

The workflow broke previously because the `prerelease` branch was
force-pushed and the old pinned commit was removed. I'm keeping the
pinned commit since this is running with the `pull_request_target` event
because of security considerations, even if it may mean it breaks again
in the future. If that happens, most likely only the commit needs to be
updated.

As before, the reporting may be too verbose, in which case we need to
silence this a bit, but that can be changed over time.