Rephrase shells (#205)
The previous version was suggesting the usage of getent, which bares the
risk of users being able to access personal data they should not see.
(although, this still being the case, we maybe don't want to advertise
such commands publicly)

Furthermore, the warning against changing the default shell was
rephrased.

---------

Co-authored-by: Ben Cumming <bcumming@cscs.ch>
Co-authored-by: Mikael Simberg <mikael.simberg@iki.fi>