Add note about missing user ids when starting containers (#282)
This seems to be a recurring issue that users bump into. Since it isn't
very easy to guess what's wrong from the error message I propose to add
this to the docs, so that the error message is at least searchable.

I've added it to the containers section now, but it could also/instead
added to the slurm section. Thoughts?

---------

Co-authored-by: Rocco Meli <r.meli@bluemail.ch>