Add `lfs migrate` examples to storage guide (#297)
This adds the `lfs migrate` recommendation/example that was removed
along with the Gordon-Bell guide (#240).

This also adds an example to recursively update all files in a directory
(finding files using `lfs find`).

@mpasserini can you or someone else from the storage team comment if you
think these are ok recommendations. For example:
- (this already exists) do you think the recommended block sizes and
progressive layouts are sane recommendations?
- Are there any concerns with using e.g. `xargs -P` to migrate files in
parallel (the example is currently sequential which can be slow, but I'm
not sure if doing a parallel `lfs migrate` adds unnecessarily high load
on lustre)?

Note that I changed the short options (`-c`/`-S`/`-E`) to long options
as I find that it's easier to follow for someone like me who doesn't
know the options by heart. However, the commands do end up being longer
(wider than the regular page, and require scrolling), and I'm happy to
change that back (it's not the important part of this PR).

I'm not attempting to move or restructure the storage guide otherwise.

Link to generated docs:
https://docs.tds.cscs.ch/297/guides/storage/#lustre-tuning