CE: Added pages with guidelines for images on Alps (#272)
As agreed in VCUE, I have updated a set of images with foundational
resources (CUDA, MPI, NCCL, NVSHMEM), deriving them from material I use
myself, and demonstrated how to run them through the CE on Alps.

The intent of this material is to offer guidelines and suggestions about
versions, building, and running of foundational components to use on
Alps, without committing to officially supported resources.

---------

Co-authored-by: Rocco Meli <r.meli@bluemail.ch>
Co-authored-by: Ben Cumming <bcumming@cscs.ch>